Delete `pome twin reset` and `pome endpoints` (F-1728, F-1721)
`twin reset` removed `.pome/<twin>.db` and `.pome-data/<twin>/<twin>.db`. For
slack, stripe, gmail and linear that was always a no-op — `TWIN_REGISTRY[*].boot`
opens `":memory:"`, so there is no file. For github the path is whatever
`GITHUB_CLONE_DB` names (`openGitHubCloneDatabase()` defaults to that env var,
and the showcases set it to `.pome/github.db`), i.e. a file the reader named
themselves, which `rm` already deletes. And `resolveStandaloneSeed` re-applies
the seed on every boot, so a reset never decided what the next start began from.
Its one other effect was deleting `.pome/twin-status.json`.

`endpoints` printed two hand-typed route arrays (github, gmail), threw for the
other three twins, and stamped every row `semantic` from nowhere — wrong for the
github routes the twin's own `fidelity.inventory.json` tiers as `shape` or
`unsupported`. The real route surface is the generated per-twin reference page
and a running twin's `/healthz` body.

The `twin` group's own description no longer advertises `reset`, and the
`twin start` unknown-twin and `--help` cases move from the deleted
`twinReset.test.ts` into `twinStart.test.ts`.
